Move transaction checks out to consensus/tx_verify.cpp.
