+++ src/miner.cpp
@@ -12,6 +12,7 @@
 
 #include "amount.h"
 #include "consensus/merkle.h"
+#include "consensus/tx_verify.h" // needed in case of no ENABLE_WALLET
 #include "hash.h"
 #include "main.h"
 #include "masternode-sync.h"
