+++ src/test/sighash_tests.cpp
@@ -4,6 +4,8 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "data/sighash.json.h"
+
+#include "consensus/tx_verify.h"
 #include "main.h"
 #include "serialize.h"
 #include "script/script.h"
